pyproject.toml
[tool.poetry]
name = "send_kakao_self"
version = "0.1.0"
description = ""
authors = ["MinDong <mdskylover@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.27.1"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
